Restructure config into a package
datacontract/config.py becomes the datacontract/config package:
resolution.py holds the ContextVar core (getenv, config_context),
settings.py holds the Config class. Import paths are unchanged since
the package __init__ re-exports the public API.
